services/base/datamodel/docker/files/datamodel/requirements.txt
aiohttp==3.7.4.post0
aniso8601==7.0.0
async-timeout==3.0.1
attrs==21.2.0
bleach==3.3.0
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
dicomgenerator==0.4.0
factory-boy==3.2.0
Faker==8.10.2
fastdiff==0.3.0
Flask==2.0.0
Flask-GraphQL==2.0.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.4.2
multidict==5.1.0
names==0.3.0
networkx==2.5.1
nibabel==3.2.1
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pathlib==1.0.1
Pillow==8.3.0
Pint==0.17
promise==2.3
psycopg2-binary==2.8.6
pycparser==2.20
pydicom==2.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
Rx==1.6.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
singledispatch==3.6.1
six==1.16.0
slackclient==2.9.3
snapshottest==0.6.0
SQLAlchemy==1.4.15
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
timeloop==1.0.2
typing-extensions==3.10.0.0
wasmer==1.0.0
wasmer-compiler-cranelift==1.0.0
webencodings==0.5.1
Werkzeug==2.0.0
yarl==1.6.3
